File Reading & Dependency Injection
Dependency injection should now work inside the project. There is one dependency
injection container which has all the types currently in the program. I also
separated the program out into a main method which runs when the program runs,
and a StaticSiteGenerator type class that is the actual logic of the program. I
did that so that I could basically just toss all my types into the DI container
and let it resolve all of them. The main method basically just reaches into the
container for this type and runs it.

I have also implemented some basic helper types for reading files and
directories, and wired them up with some basic functionality. I need to start
thinking more about how the dependency tree is going to expand outward from here
and on how I can work tests in.