Refactoring
The file reading layer has been moved into a services configuration class, so
tests don't need to be aware of how the file layer is mocked. They just need to
provide something that looks like a dictionary of file names to file contents,
and the DI can do the rest.