Refactoring to use enumerables
I decided that I wanted to use iterators to go over each item to make the code
easier to think through. When doing that, I also realized that I would lose file
context, because I was then writing out in a different loop than I was reading,
so I no longer had the file name.  Knowing I was going to need to eventually
anyway, I decided to just bite the bullet and make an interface to Markdown and
Html files. That lets me store a file path with the content, and all is right
with the world.