Introduced an integration testing base class
This base class is simply responsible for spinning up a dependency injection
context. The tests can use that context to grab types to be tested.

The base class also exposes two dictionaries, input and output.  The file system
classes read from the "input" dictionary and write to the "output" dictionary.
Tests can initialize a file system by adding things to the "input" dictionary
and check the output of file operations by reading from the "output" dictionary.

You might wish that we could use a single dictionary.  I did, too, but
apparently if you try to mutate a dictionary inconsistently while it is being
read from, it can cause strange issues, and the timing on access was conflicting
in away I didn't totally understand at the time. This is a good target for
future refactoring.