Replacement wasn't working. Now it is.
The issue is that my base class was looking to replace the target type for
conversion, not the converter. Now the search is looking at the base class of
the renderer and the renderer base replacement property is now correctly storing
the "renderer of the given type" as its target.