Fixed bug with renderer replacement
The renderer replace extension method I added wasn't actually working
properly, so it was replacing renderers that it shouldn't. I didn't
notice that anything was wrong until I happened to try to use one of
the renderers that should have been there, but weren't anymore.

The problem is that the generic types weren't being compared
properly. The code was only looking at the generic type, not the
arguments, so Renderer<Inline> was matching for
Renderer<CodeBlock>. This shouldn't happen anymore!